+++ src/ScottPlot/Plot/Plot.cs
@@ -147,7 +147,7 @@ namespace ScottPlot
             double snapHeight = GetSettings(false).YAxis.Dims.UnitsPerPx * snapDistancePixels;
 
             foreach (IDraggable draggable in GetDraggables())
-                if (draggable.IsUnderMouse(CoordinateFromPixelX(pixelX), CoordinateFromPixelY(pixelY), snapWidth, snapHeight))
+                if (draggable.IsUnderMouse(CoordinateFromPixelX((float)pixelX), CoordinateFromPixelY((float)pixelY), snapWidth, snapHeight))
                     if (draggable.DragEnabled)
                         return draggable;
